% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Title}
\usage{
add(a, b)
}
\arguments{
\item{a}{a number}

\item{b}{a number}
}
\value{
the sum of a and b 'a + b'
}
\description{
More description
}
\details{
More details
}
\examples{
add(1,2)
}
