% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare}
\alias{compare}
\title{Compare two numbers}
\usage{
compare(x, y)
}
\arguments{
\item{x}{First argument to compare.}

\item{y}{Second argument to compare.}
}
\value{
'x == y'
}
\description{
This function tests equality of two numbers and return a logic.
}
\examples{
compare(2, 3)
compare(5:10, c(5,6,7,8,9,10))
}
