% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_function.R
\name{filter}
\alias{filter}
\title{Filter rows of data frame}
\usage{
filter(x, cond1)
}
\arguments{
\item{x}{A data frame from which to filter rows}

\item{cond1}{A condition used to filter x, e.g. \code{"name of column" > 1}}
}
\value{
A subset data frame
}
\description{
Filter rows of data frame
}
\examples{
filter(iris, "Petal.Length" < 5)
filter(iris, "Species" \%in\% c("setosa", "versicolor") )

}
