% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Filter function
This is a description for the filter function.}
\usage{
filter2(data, vector)
}
\arguments{
\item{data}{Data frame for filtering}

\item{vector}{Vector for filtering}
}
\value{
The filtered data
}
\description{
Filter function
This is a description for the filter function.
}
\examples{
filter2(iris, 2)
}
