% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter3.R
\name{filter3}
\alias{filter3}
\title{Title}
\usage{
filter3(df, ind)
}
\arguments{
\item{df}{^dataframe}

\item{ind}{^logical vector}
}
\value{
row
}
\description{
Title
}
\examples{
filter3(iris,2:3)
filter3(iris,1)
}
