% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_select.R
\name{function_select_var}
\alias{function_select_var}
\title{Title}
\usage{
function_select_var(x, data)
}
\arguments{
\item{x}{variable of either characters(variables names), integers(numeric),
or boolleans}

\item{data}{A data frame}
}
\value{
selected variables of data frame
}
\description{
Title
}
\examples{
function_select_var(c("Species","Sepal.Length"),iris)
function_select_var(1:2,iris)
function_select_var(c(TRUE,TRUE,FALSE,TRUE),IRIS)
}
