% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hello_world_function.R
\name{hello_world}
\alias{hello_world}
\title{Hello World}
\usage{
hello_world(my_name = "James")
}
\arguments{
\item{my_name}{A string representing your name.}
}
\value{
The function prints a string
}
\description{
Hello World
}
\examples{
hello_world()
}
