% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniselect.R
\name{miniselect}
\alias{miniselect}
\title{Subset columns}
\usage{
miniselect(df, col)
}
\arguments{
\item{df}{This is a dataframe.}

\item{col}{This is a column of vectors.}
}
\value{

}
\description{
Subset columns
}
