% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{my_fancy_add}
\alias{my_fancy_add}
\title{my title}
\usage{
my_fancy_add(x, y)
}
\arguments{
\item{x}{numerical}

\item{y}{numerical}
}
\value{
sum of the two numerical vectors passed
}
\description{
This will be the description.
}
\details{
These will be the details.
}
\examples{
my_fancy_add(1, 2)
my_fancy_add(1:2, 2:3)
}
