% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_filter.R
\name{my_filter}
\alias{my_filter}
\title{Filter function}
\usage{
my_filter(df, ind)
}
\arguments{
\item{df}{data frame}

\item{ind}{row indices}
}
\value{
Returns the rows indicated
}
\description{
Filter function
}
\examples{
my_filter(iris, 3)
}
