% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penis_music.R
\name{my_select}
\alias{my_select}
\title{Select columns}
\usage{
my_select(df, v_names)
}
\arguments{
\item{df}{a \code{data.frame}, or another object that can be indexed with \verb{[row, col]}}

\item{v_names}{A vector of strings or integers}
}
\value{
df, but only the columns you asked for
}
\description{
Take specific columns from a data-frame
}
\examples{
my_select(iris, c("Species", "Petal.Length"))
my_select(iris, 1:4)
}
