% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrfct.R
\name{myrfct}
\alias{myrfct}
\title{Title}
\usage{
myrfct(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\description{
This will be the description.
}
\details{
These will be the details.
}
\examples{
myrfct(1,2)
myrfct(1:2,2:3)
}
