% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myselect.R
\name{myselect}
\alias{myselect}
\title{Select variables from dataframe}
\usage{
myselect(A, x)
}
\arguments{
\item{x}{a vector of either characters (variable names), integers (variable indices), or booleans.}

\item{of}{A data frame}
}
\value{
A dataframe with the subsetted variables.
}
\description{
Select variables from dataframe
}
\examples{

myselect(iris,c("Sepal.Length","Species"))
myselect(iris,c(1,2))
myselect(iris,c(TRUE,TRUE,FALSE,FALSE,FALSE))

}
