% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plustwofunction.R
\name{plustwo}
\alias{plustwo}
\title{Plus 2 Function}
\usage{
plustwo(x)
}
\arguments{
\item{x}{The number which 2 will be added to}
}
\value{
Returns x + 2
}
\description{
This function adds 2 to the input object
}
\examples{
plustwo(1)
plustwo(c(1:10))

}
