% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_function.R
\name{select}
\alias{select}
\title{Select}
\usage{
select(x, var)
}
\arguments{
\item{x}{A dataframe from which columns should be selected}

\item{var}{A vector specifying which columns should be selected from data frame \emph{x}, either indices, variable names or logicals.}
}
\value{
A data frame only consisting of the columns that have been selected.
}
\description{
Subset columns from a data frame
}
\examples{
select(iris, 1:2)
select(iris, "Sepal.Width")
select(iris, c(TRUE, FALSE, FALSE, TRUE, TRUE))
}
