% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_function.R
\name{select2}
\alias{select2}
\title{select2 - function that select columns}
\usage{
select2(data, vector)
}
\arguments{
\item{data}{data.frame}

\item{vector}{columns}
}
\value{
data.frame with selected columns
}
\description{
select2 - function that select columns
}
\examples{

select2(iris, 1:2)





}
