% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_func}
\alias{select_func}
\title{Title}
\usage{
select_func(df, inp)
}
\arguments{
\item{df}{data frame}

\item{inp}{input variables}
}
\value{
the selected variables from the data frame
}
\description{
Title
}
\examples{
select_func(iris, 1:3)
}
