% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exe1.R
\name{select_variables}
\alias{select_variables}
\title{Select Variables from a Data Frame}
\usage{
select_variables(data, vars)
}
\arguments{
\item{data}{A data frame where variables will be selected.}

\item{vars}{A character vector (for variable names) or an integer vector (for variable positions) specifying the variables to select.}
}
\value{
A data frame containing the selected variables.
}
\description{
This function selects variables from a data frame based on a character vector of variable names or an integer vector of variable positions.
}
\examples{
# Create a sample data frame
df <- data.frame(
  A = 1:5,
  B = letters[1:5],
  C = rnorm(5)
)

}
