% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{A first function}
\usage{
test(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The input \code{x} plus 1.
}
\description{
A first function that just adds 1.
}
\details{
No \emph{other} details about the \strong{function} \code{test} (or \code{test}).
Have also a look at \code{\link[=test2]{test2()}}.

You can use RMarkdown code in your description. For instance,
\itemize{
\item First item on the list.
\item Second item on the list.
}
}
\examples{
test(2:3)
}
