% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_func}
\alias{test_func}
\title{Title}
\usage{
test_func(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{a number}
}
\value{
Subtraction of \code{x - y}
}
\description{
Description
}
\details{
More detail
}
\examples{
test_func(10, 20)
test_func(20, 10)
}
