% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{zelect}
\alias{zelect}
\title{Takes A data.frame And Returns Specified Variables}
\usage{
zelect(df, variables)
}
\arguments{
\item{df}{data.frame}

\item{variables}{vector of column names, indices or booleans,
bla bla}
}
\value{
A  data.frame with the subsetted variables.
}
\description{
Takes A data.frame And Returns Specified Variables
}
\examples{
options(max.print = 10)
zelect(iris, c("Sepal.Length", "Sepal.Width"))
zelect(iris, c(1,2))
zelect(iris, c(FALSE,TRUE,FALSE,TRUE,FALSE))

}
