% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{add}
\alias{add}
\title{Add two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number to \strong{add}.}

\item{y}{Another number to add.}
}
\value{
The sum of the two numbers or \code{x + y}.
}
\description{
Add two numbers supplied to the function.
}
\examples{
add(1, 2)
add(3:4, 5:6)
}
