% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{add}
\alias{add}
\title{add two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number to \strong{add}}

\item{y}{Another number to add with the first one}
}
\value{
\code{x + y}
}
\description{
Here the function adds two numbers
}
\examples{
add(1,2)
add(3:4, 5:6)
}
