% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Filter rows of a dataframe}
\usage{
filter2(df, condition)
}
\arguments{
\item{df}{Dataframe.}

\item{condition}{}
}
\value{
Row(s) after filter applied to the provided dataframe.
}
\description{
Filters and prints the row(s) meeting a specific condition within a dataframe.
}
