% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Filter rows}
\usage{
filter2(df, ind)
}
\arguments{
\item{df}{A data frame}

\item{cond}{Indices of type integer or logical}
}
\value{
A filtered data frame
}
\description{
Filter rows of a dataframe
}
\examples{
filter2(iris,1:3)
filter2(iris,c(TRUE,FALSE))
}
