% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{filter2}
\alias{filter2}
\title{Plagiat de Filter de Deplyr
Retourne les lignes sélectionnées}
\usage{
filter2(data, selection)
}
\arguments{
\item{data}{data frame with all the data}

\item{selection}{variables to extract, character vector of variable names or
an integer vector of variable positions}
}
\value{
subset of the data
}
\description{
Plagiat de Filter de Deplyr
Retourne les lignes sélectionnées
}
\examples{
filter2(iris, c(1, 3, 4))

}
