% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Title}
\usage{
filter2(df, ind)
}
\arguments{
\item{df}{A data.frame}

\item{ind}{A cector of indices, or logical vector}
}
\value{
A data.frame with the subsetted rows
}
\description{
Title
}
\examples{
filter2(iris, 1:3)
filter2(iris, iris$Species == "setosa")
}
