% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Filter2}
\usage{
filter2(data, cond)
}
\arguments{
\item{data}{A data frame}

\item{cond}{Logical condition}
}
\value{
A data frame
}
\description{
Lets the user filter rows in data by a logical condition
}
\examples{

filter2(iris, Species == "setosa")
filter2(iris, Sepal.Length > 5.0)

}
