% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{filter2}
\usage{
filter2(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{col}{A vector of indices.}
}
\description{
Filter rows of a data frame.
}
\examples{
filter2(iris, 1:2)
}
