% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.r
\name{filter2}
\alias{filter2}
\title{Filter rows from a data.frame}
\usage{
filter2(df, ...)
}
\arguments{
\item{df}{input data.frame}

\item{...}{vector of indices or logicals}
}
\value{
data.frame with subsetted rows
}
\description{
Filter rows from a data.frame
}
