% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_function.R
\name{filter2}
\alias{filter2}
\title{Filter rows}
\usage{
filter2(.data, .rows)
}
\arguments{
\item{.data}{A data frame}

\item{.rows}{A vector of indices, variable names, or logical}
}
\value{
A data frame with the subsetted rows
}
\description{
Select rows from a data frame
}
\examples{
filter2(.data = iris, 1:3)
}
