% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{filter2}
\alias{filter2}
\title{Filter by row}
\usage{
filter2(df, rows)
}
\arguments{
\item{df}{data frame}

\item{rows}{vector}
}
\value{
a data frame with selected rows
}
\description{
Filter by row
}
\examples{
filter2(iris, 2:5)
}
