% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{filter rows}
\usage{
filter2(data, variable)
}
\arguments{
\item{data}{a data frame}

\item{variable}{use name, logical vector}
}
\value{
a dataframe with subsetted rows
}
\description{
filter rows in data frame
}
