% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Filter by row}
\usage{
filter2(data, selection)
}
\arguments{
\item{data}{a data.frame}

\item{selection}{seleted rows}
}
\value{
a filter data.frame
}
\description{
Filter by row
}
\examples{
filter2(iris, c(1:10))

}
