% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Title}
\usage{
filter2(df, rows)
}
\arguments{
\item{df}{}

\item{rows}{}
}
\value{
A data frame containing only the selected rows. The original data frame is not modified.
}
\description{
Title
}
\examples{
filter2(iris,1:10)

}
