% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter2}
\alias{filter2}
\title{Title}
\usage{
filter2(x, ind)
}
\arguments{
\item{x}{A data frame.}

\item{ind}{Either
\itemize{
\item a character vector of variable names
\item an integer vector of variables positions
\item a boolean vector of variables to keep
}}
}
\description{
Title
}
