% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter2}
\alias{filter2}
\title{Filter dataframe}
\usage{
filter2(df, flt)
}
\arguments{
\item{df}{Dataframe}

\item{flt}{Filter}
}
\value{
Filtered dataframe
}
\description{
Filter dataframe
}
\examples{
filter2(df, 1)
filter2(df, 20:40)
}
