% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Function to subset/filter rows from a data frame}
\usage{
filter2(x, rows)
}
\arguments{
\item{x}{A dataframe}

\item{rows}{The rows you want to subset}
}
\description{
Function to subset/filter rows from a data frame
}
\examples{
filter2(iris,2)
filter2(iris, 1:10)

}
