% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter2}
\alias{filter2}
\title{Filter a Data Frame}
\usage{
filter2(df, cond)
}
\arguments{
\item{df}{}

\item{cond}{}
}
\value{
filter Data Frame
}
\description{
Filter a Data Frame
}
\examples{
filter2(iris,1:3)
filter2(iris,c(TRUE,FALSE))
}
