% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Filter rows}
\usage{
filter2(df, ind)
}
\arguments{
\item{df}{A data frame}

\item{ind}{A vector of indices, or a logical vector}
}
\description{
Filter rows
}
\examples{
filter2(iris, 1:3)
filter2(iris, iris$Species == "setosa")
}
