% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Title}
\usage{
filter2(df, ind)
}
\arguments{
\item{df}{df A dataframe.}

\item{ind}{ind A vector of indices.}
}
\value{
A subsetted dataframe with less rows.
}
\description{
Title
}
\examples{
filter2(iris, 3:5)


}
