% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter2}
\alias{filter2}
\title{Filter}
\usage{
filter2(x, row_name)
}
\arguments{
\item{x}{Dataframe to filter from}

\item{row_name}{Integer or vector to filter on}
}
\description{
Filters rows in a dataframe
}
