% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_function.R
\name{filter2}
\alias{filter2}
\title{filter2 - function that filters rows}
\usage{
filter2(data, vector)
}
\arguments{
\item{data}{data.frame}

\item{vector}{rows}
}
\value{
data.frame with subset of rows
}
\description{
filter2 - function that filters rows
}
\examples{
# test

iris_test <- iris
iris_test
iris_test[1:5, ]

# use filter to select rows
filter2(iris_test, 1:5)

test_type <- filter2(iris_test, 1:5)
str(test_type)
}
