% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: A MiniPackage Reimplementing Dplyr}
\description{
A package to learn how to make package.
}
\author{
\strong{Maintainer}: Antoine Blanc \email{blanc01.antoine@gmail.com}

}
\keyword{internal}
