% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: What the Package Does (One Line, Title Case)}
\description{
This is a minidplyr package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/boztdk/minidplyr}
  \item Report bugs at \url{https://github.com/boztdk/minidplyr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
