% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: Plagiat of Real Dplyr}
\description{
This doing the plagiat of dplyr.
}
\author{
\strong{Maintainer}: Clementine Decamps \email{first.last@example.com}

}
\keyword{internal}
