% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: miniplyr: is a smaller version of dplyr}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Christoffer Trier Maansson \email{ctm@clin.au.dk} (\href{https://orcid.org/0000-0002-3071-3437}{ORCID})

}
\keyword{internal}
