% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: select2}
\description{
This package was created as a test.
}
\author{
\strong{Maintainer}: Emma Johannsen \email{emjo@clin.au.dk} (\href{https://orcid.org/0000-0002-1017-7702}{ORCID})

}
\keyword{internal}
