% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geovannazc/minidplyr}
  \item Report bugs at \url{https://github.com/geovannazc/minidplyr/issues}
}

}
\author{
\strong{Maintainer}: Geovanna Zarate \email{geovannazc@gmail.com}

}
\keyword{internal}
