% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R, R/select2.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Henrik Lynge Hovgaard \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
