% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: Sum up two numeric values}
\description{
The most awesome package in the world.
}
\author{
\strong{Maintainer}: Jesper Nors \email{jesper@jespernors.dk} (\href{https://orcid.org/0000-0002-9104-7263}{ORCID})

}
\keyword{internal}
