% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: A test package created in connection to the Course Advanced R}
\description{
This package was created in connection to the Course Advanced R that was held by Florian Prive in September 2022.
}
\author{
\strong{Maintainer}: Jette Steinbach \email{jst.ncrr@au.dk} (\href{https://orcid.org/Not_available_at_the_moment}{ORCID})

}
\keyword{internal}
