% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: Nothing}
\description{
Little Thing
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
