% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: Nothing Matters}
\description{
Little Thing. Simple Thing.
}
\author{
\strong{Maintainer}: Basile Jumentier \email{basile.jumentier@gmail.com}

}
\keyword{internal}
