% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: What the Package Does (One Line, Title Case)}
\description{
Mini version of dplyr for fun.
}
\author{
\strong{Maintainer}: Karoline Kondrup \email{kkt@clin.au.dk}

}
\keyword{internal}
