% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skrivebord-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: Simple calculations and creating new dataframes with selected variables}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lauraruthbisgaard/minidplyr}
  \item Report bugs at \url{https://github.com/lauraruthbisgaard/minidplyr/issues}
}

}
\author{
\strong{Maintainer}: Laura Bisgaard \email{lrb@anivet.au.dk}

}
\keyword{internal}
