% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiniDplyr-package.R
\docType{package}
\name{MiniDplyr-package}
\alias{MiniDplyr}
\alias{MiniDplyr-package}
\title{MiniDplyr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LeaBreniere/minidplyr}
  \item Report bugs at \url{https://github.com/LeaBreniere/minidplyr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
