% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: minidplyr}
\description{
This package is used to learn how to develop a R package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ManuThomasen/minidplyr}
  \item Report bugs at \url{https://github.com/ManuThomasen/minidplyr/issues}
}

}
\author{
\strong{Maintainer}: Manu Thomasen \email{manuth@ph.au.dk}

}
\keyword{internal}
