% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R
\docType{package}
\name{minidplyr-package}
\alias{minidplyr}
\alias{minidplyr-package}
\title{minidplyr: minidplyr}
\description{
Sum of x and y.
}
\author{
\strong{Maintainer}: Theresa Wimberley \email{tw.ncrr@au.dk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
