% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minidplyr-package.R
\docType{package}
\name{Minidplyr-package}
\alias{Minidplyr}
\alias{Minidplyr-package}
\title{Minidplyr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
