% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minidplyr-package.R
\docType{package}
\name{Minidplyr-package}
\alias{Minidplyr}
\alias{Minidplyr-package}
\title{Minidplyr: Reimplementing dplyr}
\description{
Package for learning.
}
\author{
\strong{Maintainer}: Guillaume Rivollier \email{first.last@example.com}

}
\keyword{internal}
