% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{my_fancy_add}
\alias{my_fancy_add}
\title{Title}
\usage{
my_fancy_add(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\description{
This will be the description.
}
\details{
These will be the details.
}
\examples{
my_fancy_add(1,2)
my_fancy_add(2:3,4:5)
}
