% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{my_fancy_add}
\alias{my_fancy_add}
\title{My title}
\usage{
my_fancy_add(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\value{
A numeric
}
\description{
This will automatically be the description, no need to insert "@description"
}
\details{
After another blank line: the details
}
\examples{
my_fancy_add(2, 3)
my_fancy_add(1:2, 2:3)
}
