% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_select.R
\name{my_select}
\alias{my_select}
\title{Select variables from data frame}
\usage{
my_select(df, variable_list)
}
\arguments{
\item{df, }{a data frame}

\item{variable_list, }{a character vector of variable names or an integer
vector of variable positions}
}
\value{
A data frame only with the selected variables
}
\description{
A function that helps you select variables of a data frame
}
\details{
These will be the details
}
\examples{
df_test<-data.frame(x = c(1,2,3),y = c(4,5,6),z = c(7,8,9))

my_select(df_test,c("x","y"))

my_select(df_test,c(1,3))

}
