% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_select.R
\name{my_select}
\alias{my_select}
\title{Selection of specific variables}
\usage{
my_select(df, ind)
}
\arguments{
\item{df}{insert a dataframe which includes your desired variables}

\item{ind}{the index position(s) of your desired variables, as a value, or vector}
}
\value{
Returns the selected variables
}
\description{
Select the variables desired
}
\examples{
my_select(iris, 1)
my_select(iris, 1:3)
my_select(iris, -1)
my_select(iris, c("Sepal.Length", "Species"))
my_select(iris, c(TRUE, FALSE, TRUE))
}
