% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exercise_select_ package.R, R/Select.R
\name{my_select}
\alias{my_select}
\title{Select variables}
\usage{
my_select(df, ind)

my_select(df, ind)
}
\arguments{
\item{df}{A dataframe}

\item{ind}{The index position of your desired variables,}
}
\value{
the selected variables

the selected variables
}
\description{
Select a subset of variables from a dataframe

Select a subset of variables from a dataframe
}
\details{
Selection of specific variables

Select the variables desired

Selection of specific variables

Select the variables desired
}
\examples{
my_select(iris, 1)
my_select(iris, 1:3)
my_select(iris, -1)
my_select(iris, c("Sepal.Length", "Species"))
my_select(iris, c(TRUE, FALSE, TRUE))

my_select(iris, 1)
my_select(iris, 1:3)
my_select(iris, -1)
my_select(iris, c("Sepal.Length", "Species"))
my_select(iris, c(TRUE, FALSE, TRUE))

}
