% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select variable from a dataframe}
\usage{
select(df, var)
}
\arguments{
\item{df}{Dataframe.}

\item{var}{Variable/column name(s) or number(s).}
}
\value{
Variable(s) selected from provided dataframe.
}
\description{
Selects and prints a specific variable or multiple variables of a dataframe.
}
\examples{
minidplyr::select(iris, "Species")
minidplyr::select(iris, c("Petal.Width", "Species"))
minidplyr::select(iris, 5)
minidplyr::select(iris, 4:5)
}
