% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select variable from a dataframe}
\usage{
select(df, var)
}
\arguments{
\item{df}{Dataframe.}

\item{var}{Variable/column name, provided as a character string in quotes or the number of the column.}
}
\value{
Vector of variable selected from provided dataframe.
}
\description{
Uses the {magrittr} pipe (\%>\%) and {dplyr} function "pull" to select and print a specific variable from a dataframe.
}
\examples{
minidplyr::select(iris, "Species")
minidplyr::select(iris, c("Petal.Width", "Species"))
minidplyr::select(iris, 5)
minidplyr::select(iris, 4:5)
}
