% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select}
\usage{
select(data, vars)
}
\arguments{
\item{data}{A data frame}

\item{vars}{Variables (character strings or column indices)}
}
\value{
A subsetted data frame
}
\description{
Selects variables from a data frame and returns a subsetted data frame containing only the selected variables
}
\examples{
minidplyr::select(iris, c("Petal.Width"))
minidplyr::select(iris, c(2:5))

}
