% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{select}
\alias{select}
\title{Select variables in data.frame}
\usage{
select(df, variables)
}
\arguments{
\item{df}{a data frame}

\item{variables}{either a vecotr of variable names or indexs}
}
\value{
a date frame with selected variables
}
\description{
this function is the same as select function from dplyr
}
\details{
i just make a small test for this
}
\examples{
select(iris,"Species")
select(iris,3)
}
