% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{select}
\alias{select}
\title{Subset variables from data frame}
\usage{
select(df, vars_select)
}
\arguments{
\item{df:}{A data frame}

\item{vars_select:}{A vector of either characters (variables, names), inegers (variable indices), or boolians}
}
\description{
Selects the variables of interest
}
\details{
Returns a dataframe
}
\examples{
select(palmerpenguins::penguins, "species")
select(palmerpenguins::penguins, c("species", "island"))
select(palmerpenguins::penguins, 1:2)
select(palmerpenguins::penguins, c(TRUE, FALSE, TRUE, TRUE, FALSE, TRUE, TRUE, FALSE))

}
