% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select}
\usage{
select(data, var)
}
\arguments{
\item{data}{The data frame to select from.}

\item{var}{The column to use as criteria to search.}

\item{value}{The value that should fit}
}
\value{
The rows containing a fitting variable
}
\description{
Selects columns from a data frame.
}
\examples{
select(iris, 1:2)
select(iris, "Species")
select(iris, c(T, F, F))
}
