% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectfunction.R
\name{select}
\alias{select}
\title{Subset columns of a data frame}
\usage{
select(x, var)
}
\arguments{
\item{x}{a Data Frame}

\item{var}{Either a vector of column indices or variable names or logicals}
}
\value{
A subset data frame
}
\description{
Subset columns of a data frame
}
