% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select}
\usage{
select(x, var)
}
\arguments{
\item{x}{Dataframe}

\item{var}{Either a vector of column indices or varibale names  or logicals}
}
\value{
A subset dataframe
}
\description{
Subset columns from a dataframe
}
\examples{
select(iris, 1:2)
select(iris, "Species")
select(iris, c(TRUE, FALSE))
}
