% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slctr.R
\name{select2}
\alias{select2}
\title{Simple select function using subset()}
\usage{
select2(ds, vars)
}
\arguments{
\item{ds}{Data frame}

\item{vars}{Variables to be subset, quoted or numbered}
}
\value{
Data frame with the specified subset
}
\description{
Simple parallel select function
}
\examples{
select2(iris,"Petal.Width")
}
