% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{Select variables}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{dataframe dont on veut selectionner la/les colonne(s)}

\item{ind}{indices en chaine de caracteres, integer ou logical indiquant les variables a selectionner}
}
\value{
Les colonnes selectionnees.
}
\description{
Selection des variables d'un dataframe.
}
\examples{
select2(iris,1:3)
select2(iris,c(TRUE,TRUE,TRUE,FALSE,FALSE))
select2(iris,c("Sepal.Length","Species"))
}
