% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{select2}
\alias{select2}
\title{Plagiat de Select de Deplyr
Retourne les colonnes sélectionnées}
\usage{
select2(data, selection)
}
\arguments{
\item{data}{data frame with all the data}

\item{selection}{variables to extract, character vector of variable names or
an integer vector of variable positions}
}
\value{
subset of the data
}
\description{
Plagiat de Select de Deplyr
Retourne les colonnes sélectionnées
}
\examples{
select2(iris, c(1, 3, 4))

}
