% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{select2}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{A \code{data.frame} to select columns from}

\item{ind}{Either a \code{character} or \code{numeric} vector}
}
\value{
A \code{data.frame} subsetted with \code{ind}
}
\description{
Can select
}
