% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{select2}
\usage{
select2(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{A vector of indices, variables, or logicals.}
}
\value{
A data frame with the subsetted variables.
}
\description{
select2
}
\examples{
select2(iris, 1:3)
select2(iris, 1)
select2(iris, 0)
select2(iris,"Species")
}
