% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{Select variables}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{Your dataframe}

\item{ind}{A chr or int vector}
}
\value{
A data frame with selected variables
}
\description{
Select variables
}
