% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{Subset variables from dataframe}
\usage{
select2(df, vars)
}
\arguments{
\item{df}{A dataframe}

\item{vars}{A vector of either charachters (variables names), integers (variable integers), or of booleans}
}
\value{
A dataframe with the subsetted variables
}
\description{
This is the description
}
\details{
This is the details
}
