% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.r
\name{select2}
\alias{select2}
\title{Select variables from a data.frame}
\usage{
select2(df, ...)
}
\arguments{
\item{df}{input data.frame}

\item{...}{vector of indices, variable names or logicals}
}
\value{
data.frame of length equal to length of input vector
}
\description{
Select variables from a data.frame
}
