% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{Select2}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{a data frame}

\item{ind}{a vector of variable names in the dataframe df}
}
\value{
returns the exact same thing as the function "select()" from dplyr
}
\description{
Select2
}
\examples{
select2(iris, sepal.length)

}
