% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr-package.R, R/select2.R
\name{select2}
\alias{select2}
\title{Title}
\usage{
select2(df, ind)

select2(df, ind)
}
\arguments{
\item{df}{a data frame}

\item{ind}{a vector of variable names in the dataframe df}
}
\value{
returns the exact same thing as the function "select()" from dplyr

returns the exact same thing as the function "select()" from dplyr
}
\description{
Title

Select2
}
