% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{Select variables}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{}

\item{ind}{}
}
\value{
A data frame with subsetted variables
}
\description{
Select variables
}
\examples{
select2(iris, 1:2)
}
