% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_function.R
\name{select2}
\alias{select2}
\title{Select variables}
\usage{
select2(.data, .cols)
}
\arguments{
\item{.data}{A data frame}

\item{.cols}{A vector of indices, variable names, or logical}
}
\value{
A data frame with the subsetted variables
}
\description{
Select variables from a data frame
}
\examples{
select2(.data = iris, 1:3)
}
