% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{Select column in a dataframe}
\usage{
select2(data, variable)
}
\arguments{
\item{data}{a dataframe}

\item{variable}{a variable/column in the datafram, the number or name, you can choose more than one.}
}
\value{
a datafram with the subsetted variable/variables
}
\description{
Select column in a dataframe
}
\examples{
select2[iris,"Sepal.Lenght"]
select2[iris,1]
select2[iris,1:3]

}
