% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{select2}
\alias{select2}
\title{select2}
\usage{
select2(data, selection)
}
\arguments{
\item{data}{a data.frame}

\item{selection}{interger, logical, or character vector}
}
\value{
the choosen column
}
\description{
A function to select column in a data.frame
}
\examples{
x <- data.frame(x1 = rnorm(10), x2 = rnorm(10))

select2(x, "x1")
select2(x, 1)
}
