% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{select2}
\alias{select2}
\title{Subset variables from data frame}
\usage{
select2(df, vars)
}
\arguments{
\item{df}{A data frame}

\item{vars}{A vector of either characters (variables, names), integers (variable indices), or boolian}
}
\value{
A data frame
}
\description{
Selects the variables of interest
}
\details{
Returns a dataframe
}
\examples{
select2(palmerpenguins::penguins, "species")
select2(palmerpenguins::penguins, c("species", "island"))
select2(palmerpenguins::penguins, 1:2)
select2(palmerpenguins::penguins, c(TRUE, FALSE, TRUE, TRUE, FALSE, TRUE, TRUE, FALSE))

}
