% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectfunc.R
\name{select2}
\alias{select2}
\title{Title
Select variables from a data frame}
\usage{
select2(df, vars)
}
\arguments{
\item{df}{}

\item{vars}{}
}
\value{
A data frame with only chosen variables
}
\description{
Title
Select variables from a data frame
}
