% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select2}
\alias{select2}
\title{Select variables}
\usage{
select2(x, ind)
}
\arguments{
\item{x}{A data frame.}

\item{ind}{Either
\itemize{
\item a character vector of variable names
\item an integer vector of variables positions
\item a boolean vector of variables to keep
}}
}
\value{
A data frame constituted by all the rows of \code{x} and only the columns specified in \code{ind}.
}
\description{
This function allows to select variables from a data frame
}
\examples{
select2(iris, c("Petal.Width", "Species"))
select2(iris, c(4, 5))
select2(iris, c(FALSE, FALSE, FALSE, TRUE, TRUE))
}
