% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{select2}
\alias{select2}
\title{Select 2}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{^ dataframe}

\item{ind}{^ vector of indeces variable names or logicals}
}
\value{
^ data frame with the subsetted variables
}
\description{
select variables from dataframe
}
\examples{
select2(iris,1:3)
select2(iris,1)
select2(iris,0)

}
