% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select2}
\alias{select2}
\title{Select some columns}
\usage{
select2(df, names_or_ind)
}
\arguments{
\item{df}{A data frame}

\item{names_or_ind}{A vector of column names or indices}
}
\value{
An other data frame, with the subsetted columns.
}
\description{
Select some columns
}
