% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{Select Variables from a Data Frame}
\usage{
select2(data, vars)
}
\arguments{
\item{data}{A data frame from which variables will be selected.}

\item{vars}{A character vector of variable names or an integer vector
of column positions.}
}
\value{
A data frame containing only the selected variables.
}
\description{
This function selects variables (columns) from a data frame based on
either a character vector of variable names or an integer vector of
variable positions. It is a simple wrapper around the standard
subsetting operator \code{[} for data frames.
}
\examples{
df <- data.frame(a = 1:5, b = 6:10, c = 11:15)

# Select by name
select2(df, c("a", "c"))

# Select by position
select2(df, c(1, 3))

}
