% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{Select Specific Names or Idx}
\usage{
select2(df, names_or_idx)
}
\arguments{
\item{df}{A data frame or tibble object}

\item{names_or_idx}{A vector of column names or column indices.}
}
\value{
Data frame, with the specified columns subset
}
\description{
Select based on specific column names or idx
}
\details{
Any additional description \strong{Here}
}
