% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select2}
\alias{select2}
\title{Select variables of a data frame}
\usage{
select2(x, vars)
}
\arguments{
\item{x}{a data frame}

\item{vars}{Either
\itemize{
\item a variable character vector of variable names
\item a integer vector of variable positions
\item a boolean vector of variables to keep
}}
}
\value{
A data frame with the selected variables
}
\description{
Select variables of a data frame
}
\examples{
select2(iris,2)
select2(iris,"Species")
select2(iris,c(TRUE,FALSE,TRUE,FALSE,FALSE))
}
