% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select2}
\alias{select2}
\title{Select index from dataframe}
\usage{
select2(df, indx)
}
\arguments{
\item{df}{Dataframe}

\item{indx}{Index, either a \emph{character}, \emph{boolean} or \emph{integer} vector}
}
\value{
A dataframe with index \code{indx} of the dataframe \code{df}
}
\description{
The function \code{select2} selects an index (\code{indx}) of a dataframe (\code{df})
using a \emph{character}, \emph{boolean} or \emph{integer} vector
}
\examples{
select2(iris, c(2,4))
select2(iris, c("Sepal.Length","Species"))
select2(iris, c(TRUE, TRUE, FALSE, FALSE, TRUE))

}
