% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{select variables}
\usage{
select2(df, vars)
}
\arguments{
\item{df}{a dataframe}

\item{cols_to_select}{character vector of variable names or integer for index}
}
\value{
returns data frame with selected column(s)
}
\description{
This function selects columns
}
\examples{
select2(df = iris, cols_to_select = "Species")
select2(df = iris, vars = 5)


}
