% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{Subset variables from data frame}
\usage{
select2(df, vars)
}
\arguments{
\item{df}{A data frame}

\item{vars}{A vector of either characters (variable names),
integers (variable indices), or booleans.}
}
\value{
A data frame with the subsetted variables.
}
\description{
A function that helps you select variables of a data frame
}
\examples{
select2(iris, c("Sepal.Length", "Species"))
select2(iris, 1:2)
select2(iris, c(TRUE, TRUE, FALSE, FALSE, FALSE))
}
