% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select2}
\alias{select2}
\title{Select}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{A data frame}

\item{ind}{A vector of indicies, variable names or logicals.}
}
\description{
This is the description of select
}
\examples{

select2(iris, "Species")

}
