% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select2}
\alias{select2}
\title{Select variables}
\usage{
select2(x, ind)
}
\arguments{
\item{x}{A data frame.}

\item{ind}{Either
\itemize{
\item a character vector of variable names
\item an integer vector of variable positions
\item a boolean vector of variables to keep
}}
}
\value{
A data frame with only the selected variables.
}
\description{
Select variables from a data frame
}
\examples{
select2(iris, c("Sepal.Length", "Sepal.Width"))
select2(iris, 1:4)
select2(iris, c(TRUE, FALSE, TRUE, FALSE, TRUE))
}
