% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select2}
\alias{select2}
\title{Select variable}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{A data frame}

\item{ind}{Indices as integer, logical or character vector.}
}
\value{
A subsetted data frame
}
\description{
Select variable from a data frame
}
\examples{
select2(iris, 1:3)
select2(iris, c(TRUE, TRUE, TRUE, FALSE, FALSE))
}
