% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select2}
\alias{select2}
\title{Select2}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{A data frame.}

\item{ind}{A vector of indices, variables names, or logicals.}
}
\value{
A data frame with the subsetted variables.
}
\description{
Select variables
}
\details{
Select variables from a data frame.
}
\examples{
select2(iris, 1:3)
select2(iris, 1)
select2(iris, 0)
select2(iris, "Species")
}
