% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select2}
\alias{select2}
\title{Select variable}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{A data frame}

\item{ind}{A vector of indicies, variable names, or logicals}
}
\value{
A data frame with subsetted variables
}
\description{
A function used to choose whether a column of a data frame is selected or not.
}
\examples{
select2(iris, 1:3)
select2(iris, "Species")
}
