% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select2}
\alias{select2}
\title{Select columns from dataframe}
\usage{
select2(df, names_or_ind)
}
\arguments{
\item{df}{Dataframe to select from}

\item{names_or_ind}{A vector of column names or numbers}
}
\value{
another data frame
}
\description{
Select columns from dataframe
}
