% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{Title}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{A dataframe}

\item{ind}{A vector of indices, or variable name}
}
\value{
A daraframe with selected columns
}
\description{
Title
}
\examples{
select2(iris, 1:2)
}
