% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Adding two numbers}
\usage{
test(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of x and y
}
\description{
This is the description
}
\details{
This will be more description
}
\examples{
test(2, 5)
test(10, 16)

}
