% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{A first test function}
\usage{
test(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The input 'x' plus 1.
}
\description{
here the content will go to description
}
\details{
here will go to details \code{test}
}
\examples{
test(2:3)

}
