% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{My first function}
\usage{
test(x, y)
}
\arguments{
\item{x}{number1 or vector or data frame}

\item{y}{number2 or vector or data frame}
}
\value{
The sum of x and y
}
\description{
Sum to variables
Not so much to explain...
}
\examples{
test(4, 5)
test(1:4, 3:6)
}
