% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Title}
\usage{
test(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum `x + y´
}
\description{
This will be the description
}
\details{
This will be more details
}
\examples{

test (4,5)
test(3:4, 7:8)
}
